Change the AdvancedScanner interface to use ColumnNames() instead of ColumnTypes()
ColumnTypes returns ([]*sql.ColumnType, error). This makes it impossible for any
other data sources to make wrappers or implement their own column type lists
without going through a lot of effort to mock their own database driver because
sql.ColumnTypes is a solid struct with no exported fields, so it's impossible to
build your own lists of columns.

So, replace it with ColumnNames(), which just returns []string. It's easy to map
from one to the other.