Rewrite cache.go.
Untangle some of the messier parts of the logic and fix some of the
badly named functions and structs.

It works like this now:
At the lowest level, there is fieldCache (and fieldCacheLock), and
buildFieldCacheEntryForType.

Just above that is getFieldCachesFor which interacts with both the
cache and the function to directly create, as needed. it also handles
locking the cache.

Adjacent to that is NamedFieldsFromGetFields, a shim which just wraps
a GetFields implementor in an implementation of NamedFields so it
satisfies the same interface as cache entries from the leval below it.

Just above that is GetNamedFieldsMakers, which returns a mix of both
getFieldCachesFor and NamedFieldsFromGetFields objects depending on
its arguments.

Prior to (and adjacent to) that is NormalizeObjects, which inspects
and prepares an incoming argument list into an appropriate set of
reflect types and values.

NormalizeObjects falls back on an internal implementation with a
special mode flag capable of allowing it to accept objects even if it
cannot write to them, which is useful when building NamedFieldsMaker
lists for later evaluation.

After (and adjacent to) that is RenderNamedFields, which takes a list
of NamedFieldsMaker and a list of object and zips them together,
merging all of the NamedFields outputs together into one result.